Refactor token graph
